services/auth/pyproject.toml
[project]
name = "az-jina-auth"
version = "0.1.0"
description = "Azure Key Vault-backed auth service for Jina Reader/MCP"
requires-python = ">=3.13"
dependencies = [
  "azure-identity>=1.17.1",
  "azure-keyvault-secrets>=4.9.0",
  "fastapi>=0.115.8",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.7.1",
  "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["auth_service"]
